pyproject.toml
[tool.poetry]
name = "modelserver"
version = "0.1.0"
description = ""
authors = ["Logan Henriquez <12344+henriquez@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
chromadb = "^0.3.7"
openai = "0.27.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
